Module Parsley

Parsley library.

This module provides the basic conversion operations over the built-in numeric types of OCaml (integers, Int32.t, Int64.t, Nativeint.t and floats) while indicating if the conversion was exact or approximated. It does it using the Result module of the standard library.

of_string conversions

val exact_float_of_string : string -> (float, float option) Stdlib.Result.t

Converts the given string to a float, following OCaml's lexical conventions. It returns Error (None) if the given string is not a valid representation of a float, or Error (Some f) if f is an approximated representation of the given string. It returns Ok f if f is an exact representation of the given string.

integers-float conversions

val exact_float_of_int : int -> (float, float) Stdlib.Result.t

Converts the given integer to a float. It returns Error f if f is an approximated representation of the given integer. It returns Ok f if f is an exact representation of the given integer.

val exact_int_of_float : float -> (int, int) Stdlib.Result.t

Converts the given foat to an integer. It returns Error i if i is an approximated representation of the given float. It returns Ok i if i is an exact representation of the given float.

val exact_32_of_float : float -> (Stdlib.Int32.t, Stdlib.Int32.t) Stdlib.Result.t

val exact_float_of_64 : Stdlib.Int64.t -> (float, float) Stdlib.Result.t

val exact_64_of_float : float -> (Stdlib.Int64.t, Stdlib.Int64.t) Stdlib.Result.t

val exact_float_of_native : Stdlib.Nativeint.t -> (float, float) Stdlib.Result.t

val exact_native_of_float : float -> (Stdlib.Nativeint.t, Stdlib.Nativeint.t) Stdlib.Result.t

cross-integers conversions

val exact_int_of_32 : Stdlib.Int32.t -> (int, int) Stdlib.Result.t

val exact_32_of_int : int -> (Stdlib.Int32.t, Stdlib.Int32.t) Stdlib.Result.t

val exact_int_of_64 : Stdlib.Int64.t -> (int, int) Stdlib.Result.t

val exact_64_of_int : int -> (Stdlib.Int64.t, Stdlib.Int64.t) Stdlib.Result.t

val exact_int_of_nativeint : Stdlib.Nativeint.t -> (int, int) Stdlib.Result.t

val exact_nativeint_of_int : int -> (Stdlib.Nativeint.t, Stdlib.Nativeint.t) Stdlib.Result.t

val exact_32_of_64 : Stdlib.Int64.t -> (Stdlib.Int32.t, Stdlib.Int32.t) Stdlib.Result.t

val exact_64_of_32 : Stdlib.Int32.t -> (Stdlib.Int64.t, Stdlib.Int64.t) Stdlib.Result.t

val exact_native_of_64 : Stdlib.Int64.t -> (Stdlib.Nativeint.t, Stdlib.Nativeint.t) Stdlib.Result.t

val exact_64_of_native : Stdlib.Nativeint.t -> (Stdlib.Int64.t, Stdlib.Int64.t) Stdlib.Result.t

val exact_native_of_32 : Stdlib.Int32.t -> (Stdlib.Nativeint.t, Stdlib.Nativeint.t) Stdlib.Result.t

val exact_32_of_native : Stdlib.Nativeint.t -> (Stdlib.Int32.t, Stdlib.Int32.t) Stdlib.Result.t

to_string conversions

val exact_string_of_float : float -> string

exact string_of_float: same as Stdlib.string_of_float but computes all the non-nul decimals of a floating-point number